requirements.txt
streamlit
pymongo
pandas
plotly
python-dotenv
certifi
google-generativeai